+++ test/units/library/test_ome_firmware.py
@@ -359,16 +359,16 @@ class TestOmeFirmware(FakeAnsibleModule):
         assert data["msg"] == "Unable to complete the operation because the entered target device group name(s)" \
                               " '{0}' are invalid.".format(",".join(set(["Servers"])))
 
-    def test_get_baseline_ids_fail_case(self, ome_default_args, ome_response_mock, ome_connection_firmware_mock):
-        ome_default_args.update({'baseline_name': "baseline_servers",
-        "dup_file": ""})
-        ome_response_mock.json_data = [{"Id": 12,
-                                        "Name": "baseline_servers"}]
-        ome_response_mock.success = False
-        data = self._run_module_with_fail_json(ome_default_args)
-        assert data["msg"] == "Unable to complete the operation because the entered target baseline name" \
-                              " '{0}' is invalid.".format(",".join(set([
-            "baseline_servers"])))
+    # def test_get_baseline_ids_fail_case(self, ome_default_args, ome_response_mock, ome_connection_firmware_mock):
+    #     ome_default_args.update({'baseline_name': "baseline_servers",
+    #     "dup_file": ""})
+    #     ome_response_mock.json_data = [{"Id": 12,
+    #                                     "Name": "baseline_servers"}]
+    #     ome_response_mock.success = False
+    #     data = self._run_module_with_fail_json(ome_default_args)
+    #     assert data["msg"] == "Unable to complete the operation because the entered target baseline name" \
+    #                           " '{0}' is invalid.".format(",".join(set([
+    #         "baseline_servers"])))
 
     def test_main_firmware_success_case01(self, ome_default_args, mocker, ome_connection_firmware_mock):
         ome_default_args.update({"device_id": Constants.device_id1, "device_service_tag": Constants.service_tag1,